feat: proposal for shortcuts (#69)
### Description

This is a proposal for adding support for shortcut hooks.

### Details

This is focusing on applying shortcuts to positions. But it could be
extended to support shortcuts for tokens or more generic actions.

It would support the following use cases
- claiming Good Dollar UBI
- claiming Ubeswap farm staking rewards

The shortcuts are defined in a new file under
`src/apps/{appId}/shortcuts.ts`. This is to give builders clear
visibility into what type of hooks are supported for a given app.
The plan is to rename `plugin.ts` to `positions.ts` to follow with this
new structure.

The positions are extended with a new `availableShortcutIds` property
which allows defining which shortcuts can be applied to them. Underlying
tokens get a new optional `category` property which allows giving more
meaning to what they are.
Specifically here, the `claimable` category tells the wallet the tokens
and amounts which can be claimed.

The wallet would fetch the shortcuts with a new route in `hooks-api`.
(Or we could also return both positions and shortcuts data in a single
call).

Using the new data in the returned positions, the wallet would be able
to fill the content of a new "Claim Rewards" screen.
Displaying details for the claimable tokens.
Tapping the button next to each of them would call the shortcut's
`onTrigger` function, via a new route in `hooks-api`.
It would return 0, 1 or more transaction to sign.

The wallet would take it from there and display the usual transaction
signing bottom sheet. Similar to what's happening when interacting with
dapps with WalletConnect.

### Future considerations

This could be extended to support other types of shortcuts related to
positions and requiring additional input parameters. For instance
adding/removing liquidity in Ubeswap pools. This would require a new
wallet UI to specify the amounts needed.

We could maybe think of pre-built UI elements that could be used and
defined by shortcuts so we don't have to build custom UIs in the wallet
to support them.

### Related linear task

Fixes RET-732