feat!: make `getAppTokenDefinition` optional (#197)
`getAppTokenDefinition` is only needed in 2 situations:
1. when the token definition returns intermediary tokens which are not
base tokens
2. when a hook depends on another hook's tokens which are not base
tokens

Though 2 is not yet fully supported as we haven't had the case yet, but
it exists. For instance Revo tokens depending on Ubeswap LP tokens.

So far only Ubeswap and Curve really needed to implement it.
This change aims to make the common use case simpler to implement for
hooks author.

The runtime will throw a useful error instead when
`getAppTokenDefinition` is needed but no implemented.