feat: Add Beefy Vault tracking (#396)
Updated PR description and screenshot, striking through old text.

Added tracking for Beefy on Celo, Ethereum, Optimism and Arbitrum. 

Didn't test on Celo because there are no active vaults and didn't test
in Ethereum because it's too expensive to do so with such small values.
The logic should be the same in all chains though, and several different
types of vaults have been tested on Optimism and Arbitrum.


![IMG_79801E2E5CAF-1](https://github.com/valora-inc/hooks/assets/6062888/a719f682-5227-4cc1-9f1e-77679844503e)

~~Vaults on Beefy can have different "strategies". This menas the
underlying token can be things from an LP token, a lending token from a
protocol like Aave or more complex things like concetrated liquidity
positions, GMX tokens which is a leverage-powered platform or other
types of tokens I haven't fully investigated yet.~~

~~This PR adds explicit support for the most popular type of underlying
token which is an LP token and it also works for lending tokens assuming
the underlying token is a known token like USDC. More work will be
needed to add full support for all types of underlying tokens.~~

~~On an unrelated point, I'm not sure I implemented the position
definition as intended by the interfaces. Even though there is an
underlying app token I had trouble implementing it using an
`AppTokenPositionDefinition` and used a `ContractPositionDefinition`
instead. The code that determines the balance and USD price for app
tokens is here:
https://github.com/valora-inc/hooks/blob/main/src/runtime/getPositions.ts#L171
there is a TODO there that might be related but I couldn't figure out
how it was intended to work, the `pricePerShare` and calling of
`balanceOf` confuses me a bit. I could use some help if the usage of app
token definitions is prefered.~~

In the UI side, one important thing that we should probably show somehow
is if a vault is "Retired". They create new vaults and retire old ones
somewhat frequently, so it could be easy for a user not to realize that
a vault they have their money in is retired and no longer giving out
rewards. The UI also doesn't show which network each position belongs
to.

One more thing to note is that the balance values are different on
Valora's UI than on Beefy's Dashboard page. Valora's numbers match the
value shown on-chain by calling `balanceOf` to the Beefy contracts. For
the USDC Vault for example, they seem to be doing something special with
decimals, because the on-chain contract says it has 18 decimals but they
show it as if it had just 6 in their UI (base USDC has 6 decimals as
well). The exact value for the balance is also slightly off. The total
USD price seems to match though.

![Screen Shot 2024-06-18 at 02 32
12](https://github.com/valora-inc/hooks/assets/6062888/bb554791-e5fc-4394-b960-cc50199a9b7a)