feat: Add Beefy CLM Vaults support (#423)
Beefy recently created CLM (Cowcentraded Liquidity Manager) Vaults which
work a bit differently so they aren't included in the main `/vaults`
endpoint. They can be fetched from a similar endpoint which is not
documented (they suggested I use it on Discord) called
`/harvestable-vaults` which includes all the "normal" vaults plus the
CLM ones.

They need to be handled a bit differently, since the token is an NFT,
not an ERC20, so we use a `ContractPositionDefinition` instead.