feat: optimize RPC calls by enabling client batching via multicall (#474)
I noticed getting positions without a user address could be quite slow
or return 429 errors.
This is because the runtime was doing many more RPC calls in that case
compared to when a user address is provided.

Specifically, getting earn positions was problematic on Arbitrum.
Or getting all curve positions.

On my local machine, running `yarn getPositions --networkId arbitrum-one
--apps aave` would easily take **~40 secs** (and often was returning 429
with the defaut Arbitrum RPC).
Now with this change, it takes **~4** secs, and no more 429!
Yeah **10x** speed improvement! 😄 

Note: in some hooks, we could benefit from switching to implicit
multicall via batching too. Because we fetch definitions in
[parallel](https://github.com/valora-inc/hooks/blob/a8c099d0734f9b3ba285c638c354f1a87e9e5141/src/runtime/getPositions.ts#L406-L408),
if the hook itself would call `readContract` (or via `Promise.all`), all
these async calls would be batched too.
We can try this further optimization in the future.