feat: consistent args for shortcuts (#514)
For clients to call deposit / withdraw / claim shorts for different
providers, we need consistent input parameters for different providers.
This updates position definitions to define a shortcutTriggerArgs that
can be passed as is to `triggerShortcut`.
Additionally, enforces `tokens` field on deposit / withdraw shortcuts
which is an array of tokenId's and amounts, which will be set by the
client. The aave and allbridge deposition hooks also need token address
and decimals, which can ideally be looked up using tokenId, but as a
workaround for now, is passed via `shortcutTriggerArgs`. Eventually we
can replace this with the lookup and drop the trigger args with no
impact to the client