fix: valora app version check for Aave positions (#680)
I was randomly reading the code and noticed the check to include/exclude
Aave positions would fail once we hit version `1.100`. Which will be the
next release.

This fixes it, and also defaults to including Aave for other Mobile
Stack apps (any other User-Agent).
Which is what I think we want.